requirements.txt
Django=3.0.9
djangorestframework=3.11.0
psycopg2=2.8.5
Pillow=2.2.1

flake8=3.6.0
